# back/docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.ConfiguracioEntrenador:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      duracioSessions:
        type: integer
      entrenadorID:
        type: integer
      id:
        type: integer
      maxAlumnesPerSessio:
        type: integer
      updatedAt:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.HorarisEntrenador:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      desde:
        type: string
      diaSetmana:
        type: integer
      entrenadorID:
        type: integer
      fins:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
    type: object
  models.Reserva:
    properties:
      confirmada:
        type: boolean
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      entrenadorID:
        type: integer
      hora:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
      usuariID:
        type: integer
    type: object
...
  models.Sala:
    properties:
      admin:
        $ref: '#/definitions/models.Usuari'
      adminID:
        type: integer
      codiSala:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      nom:
        type: string
      reserves:
        items:
          $ref: '#/definitions/models.Reserva'
        type: array
      updatedAt:
        type: string
      usuaris:
        items:
          $ref: '#/definitions/models.Usuari'
        type: array
    type: object
...
  models.SolicitudUnioEntrenador:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      entrenadorID:
        type: integer
      id:
        type: integer
      updatedAt:
        type: string
      usuariID:
        type: integer
    type: object
...
  models.SuccessResponse:
    properties:
      data: {}
    type: object
  models.TipusUsuari:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      nom:
        type: string
      updatedAt:
        type: string
    type: object
  models.Usuari:
    properties:
      alumnes:
        items:
          $ref: '#/definitions/models.Usuari'
        type: array
      codiEntrenador:
        type: string
      configuracioEntrenador:
        $ref: '#/definitions/models.ConfiguracioEntrenador'
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      entrenador:
        $ref: '#/definitions/models.Usuari'
      entrenadorID:
        type: integer
      horarisEntrenador:
        items:
          $ref: '#/definitions/models.HorarisEntrenador'
        type: array
      id:
        type: integer
      nom:
        type: string
      password:
        type: string
      reserves:
        items:
          $ref: '#/definitions/models.Reserva'
        type: array
      sales:
        items:
          $ref: '#/definitions/models.Sala'
        type: array
      salesAdministrades:
        items:
          $ref: '#/definitions/models.Sala'
        type: array
      solicitudsUnioEntrenador:
        items:
          $ref: '#/definitions/models.SolicitudUnioEntrenador'
        type: array
      telefon:
        type: string
      tipusUsuari:
        allOf:
        - $ref: '#/definitions/models.TipusUsuari'
        description: 1 = administrador, 2 = basic , 3 = entrenador, 4 = fictici
      tipusUsuariID:
        type: integer
      updatedAt:
        type: string
    type: object
...
  models.UsuariInput:
    properties:
      nom:
        type: string
      password:
        type: string
      tipusUsuariID:
        type: integer
    type: object
  models.UsuariResultatExerciciInput:
    properties:
      dia:
        type: string
      exerciciRutinaID:
        type: integer
      pes:
        type: number
      repeticions:
        type: integer
      series:
        type: integer
      usuariRutinaID:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/entrenador/guardarResultats:
    post:
      consumes:
      - application/json
      description: Saves the results of a day of the current routine in the database.
      parameters:
      - description: ID of the routine to save the results of
        in: path
        name: id
        required: true
        type: integer
      - description: Results to save
        in: body
        name: input
        required: true
        schema:
          items:
            $ref: '#/definitions/models.UsuariResultatExerciciInput'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Save the results of a day of the current routine
      tags:
      - Exercises
...
  /api/users:
    get:
      consumes:
      - application/json
      description: Retrieves all the users from the database.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Usuari'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Creates a new user in the database.
      parameters:
      - description: User to create
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UsuariInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Usuari'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new user
      tags:
      - Users
  /api/users/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a user from the database.
      parameters:
      - description: ID of the user to delete
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete a user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Retrieves a user from the database by its ID.
      parameters:
      - description: ID of the user to retrieve
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Usuari'
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get a user by ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Updates a user in the database.
      parameters:
      - description: ID of the user to update
        in: path
        name: id
        required: true
        type: integer
      - description: User to update
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UsuariInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Usuari'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update a user
      tags:
      - Users
securityDefinitions:
  Bearer:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"